Changed initial request command size and response status size from 1 to 4 bytes

+++ iggy/src/binary/binary_client.rs
@@ -4,5 +4,5 @@ use async_trait::async_trait;
 
 #[async_trait]
 pub trait BinaryClient: Client {
-    async fn send_with_response(&self, command: u8, payload: &[u8]) -> Result<Vec<u8>, Error>;
+    async fn send_with_response(&self, command: u32, payload: &[u8]) -> Result<Vec<u8>, Error>;
 }
